ml_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0

# ML/DL libraries (optional, uncomment when needed)
# tensorflow==2.14.0
# torch==2.1.0
# torchvision==0.16.0
# numpy==1.24.3
# opencv-python==4.8.1.78
# scikit-learn==1.3.2

